Adding X-Axis .STL files
Uploading some .STLs. Included are:

X-Axis 2020 Roller Carriage - This is the roller carriage that slides along a 2020 extrusion. Two of these are needed for the X-Axis.

X-Axis 2020 Carriage Body - This is the primary body that the roller carriages attach to. The GT2 belt for the X-Axis attaches to this component. 

X-Axis 2020 Mounting Plate - This is the top plate that attaches to the carriage body. It provides a number of holes compatible with M3 nuts/bolts for mounting things.

200mm 2020 Rail.stl - This is an example 2020 rail used for modeling/reference purposes. It is 200mm long.

400mm 2020 Rail.stl - This is an example 2020 rail used for modeling/reference purposes. It is 400mm long.